+++ src/components/safe_browsing/content/browser/web_api_handshake_checker.cc
@@ -1,12 +1,14 @@
-// Copyright 2021 The Chromium Authors. All rights reserved.
+// Copyright 2021 The Chromium Authors
 // Use of this source code is governed by a BSD-style license that can be
 // found in the LICENSE file.
 
 #include "components/safe_browsing/content/browser/web_api_handshake_checker.h"
 
+#include "base/metrics/histogram_functions.h"
 #include "components/safe_browsing/content/browser/web_ui/safe_browsing_ui.h"
 #include "components/safe_browsing/core/browser/safe_browsing_url_checker_impl.h"
 #include "components/safe_browsing/core/browser/url_checker_delegate.h"
+#include "components/safe_browsing/core/common/features.h"
 #include "content/public/browser/browser_task_traits.h"
 #include "content/public/browser/browser_thread.h"
 #include "content/public/browser/web_contents.h"
@@ -14,10 +16,10 @@
 
 namespace safe_browsing {
 
-class WebApiHandshakeChecker::CheckerOnIO
-    : public base::SupportsWeakPtr<WebApiHandshakeChecker::CheckerOnIO> {
+class WebApiHandshakeChecker::CheckerOnSB
+    : public base::SupportsWeakPtr<WebApiHandshakeChecker::CheckerOnSB> {
  public:
-  CheckerOnIO(base::WeakPtr<WebApiHandshakeChecker> handshake_checker,
+  CheckerOnSB(base::WeakPtr<WebApiHandshakeChecker> handshake_checker,
               GetDelegateCallback delegate_getter,
               const GetWebContentsCallback& web_contents_getter,
               int frame_tree_node_id)
@@ -37,7 +39,9 @@ class WebApiHandshakeChecker::CheckerOnIO
   }
 
   void Check(const GURL& url) {
-    DCHECK_CURRENTLY_ON(content::BrowserThread::IO);
+    DCHECK_CURRENTLY_ON(base::FeatureList::IsEnabled(kSafeBrowsingOnUIThread)
+                            ? content::BrowserThread::UI
+                            : content::BrowserThread::IO);
     DCHECK(delegate_getter_);
     DCHECK(web_contents_getter_);
 
@@ -48,28 +52,44 @@ class WebApiHandshakeChecker::CheckerOnIO
         url_checker_delegate->ShouldSkipRequestCheck(
             url, frame_tree_node_id_,
             /*render_process_id=*/content::ChildProcessHost::kInvalidUniqueID,
-            /*render_frame_id=*/MSG_ROUTING_NONE,
+            /*render_frame_token=*/std::nullopt,
             /*originated_from_service_worker=*/false);
     if (skip_checks) {
-      OnCompleteCheck(/*slow_check=*/false, /*proceed=*/true,
-                      /*showed_interstitial=*/false);
+      OnCompleteCheckInternal(/*proceed=*/true);
       return;
     }
 
+    // If |kSafeBrowsingSkipSubresources2| is enabled, skip Safe Browsing checks
+    // for WebTransport.
+    if (base::FeatureList::IsEnabled(kSafeBrowsingSkipSubresources2)) {
+      base::UmaHistogramBoolean("SafeBrowsing.WebApiHandshakeCheck.Skipped",
+                                true);
+      OnCompleteCheckInternal(/*proceed=*/true);
+      return;
+    }
+
+    base::UmaHistogramBoolean("SafeBrowsing.WebApiHandshakeCheck.Skipped",
+                              false);
     url_checker_ = std::make_unique<SafeBrowsingUrlCheckerImpl>(
         net::HttpRequestHeaders(), /*load_flags=*/0,
         network::mojom::RequestDestination::kEmpty, /*has_user_gesture=*/false,
-        url_checker_delegate, web_contents_getter_,
+        url_checker_delegate, web_contents_getter_, /*weak_web_state=*/nullptr,
         /*render_process_id=*/content::ChildProcessHost::kInvalidUniqueID,
-        /*render_frame_id=*/MSG_ROUTING_NONE, frame_tree_node_id_,
-        /*real_time_lookup_enabled=*/false,
-        /*can_rt_check_subresource_url=*/false,
-        /*can_check_db=*/true, last_committed_url_,
+        /*render_frame_token=*/std::nullopt, frame_tree_node_id_,
+        /*url_real_time_lookup_enabled=*/false,
+        /*can_urt_check_subresource_url=*/false,
+        /*can_check_db=*/true, /*can_check_high_confidence_allowlist=*/true,
+        /*url_lookup_service_metric_suffix=*/".None", last_committed_url_,
         content::GetUIThreadTaskRunner({}),
-        /*url_lookup_service=*/nullptr, WebUIInfoSingleton::GetInstance());
+        /*url_lookup_service=*/nullptr, WebUIInfoSingleton::GetInstance(),
+        /*hash_realtime_service_on_ui=*/nullptr,
+        /*mechanism_experimenter=*/nullptr,
+        /*is_mechanism_experiment_allowed=*/false,
+        /*hash_realtime_selection=*/
+        hash_realtime_utils::HashRealTimeSelection::kNone);
     url_checker_->CheckUrl(
         url, "GET",
-        base::BindOnce(&WebApiHandshakeChecker::CheckerOnIO::OnCheckUrlResult,
+        base::BindOnce(&WebApiHandshakeChecker::CheckerOnSB::OnCheckUrlResult,
                        base::Unretained(this)));
   }
 
@@ -78,26 +98,40 @@ class WebApiHandshakeChecker::CheckerOnIO
   void OnCheckUrlResult(
       SafeBrowsingUrlCheckerImpl::NativeUrlCheckNotifier* slow_check_notifier,
       bool proceed,
-      bool showed_interstitial) {
-    DCHECK_CURRENTLY_ON(content::BrowserThread::IO);
+      bool showed_interstitial,
+      SafeBrowsingUrlCheckerImpl::PerformedCheck performed_check) {
+    DCHECK_CURRENTLY_ON(base::FeatureList::IsEnabled(kSafeBrowsingOnUIThread)
+                            ? content::BrowserThread::UI
+                            : content::BrowserThread::IO);
     if (!slow_check_notifier) {
-      OnCompleteCheck(/*slow_check=*/false, proceed, showed_interstitial);
+      OnCompleteCheckInternal(proceed);
       return;
     }
 
     *slow_check_notifier =
-        base::BindOnce(&WebApiHandshakeChecker::CheckerOnIO::OnCompleteCheck,
+        base::BindOnce(&WebApiHandshakeChecker::CheckerOnSB::OnCompleteCheck,
                        base::Unretained(this), /*slow_check=*/true);
   }
 
-  void OnCompleteCheck(bool slow_check,
-                       bool proceed,
-                       bool showed_interstitial) {
-    DCHECK_CURRENTLY_ON(content::BrowserThread::IO);
-    content::GetUIThreadTaskRunner({})->PostTask(
-        FROM_HERE, base::BindOnce(&WebApiHandshakeChecker::OnCompleteCheck,
-                                  handshake_checker_, slow_check, proceed,
-                                  showed_interstitial));
+  void OnCompleteCheck(
+      bool slow_check,
+      bool proceed,
+      bool showed_interstitial,
+      SafeBrowsingUrlCheckerImpl::PerformedCheck performed_check) {
+    OnCompleteCheckInternal(proceed);
+  }
+
+  void OnCompleteCheckInternal(bool proceed) {
+    DCHECK_CURRENTLY_ON(base::FeatureList::IsEnabled(kSafeBrowsingOnUIThread)
+                            ? content::BrowserThread::UI
+                            : content::BrowserThread::IO);
+    if (base::FeatureList::IsEnabled(kSafeBrowsingOnUIThread)) {
+      handshake_checker_->OnCompleteCheck(proceed);
+    } else {
+      content::GetUIThreadTaskRunner({})->PostTask(
+          FROM_HERE, base::BindOnce(&WebApiHandshakeChecker::OnCompleteCheck,
+                                    handshake_checker_, proceed));
+    }
   }
 
   base::WeakPtr<WebApiHandshakeChecker> handshake_checker_;
@@ -113,29 +147,33 @@ WebApiHandshakeChecker::WebApiHandshakeChecker(
     const GetWebContentsCallback& web_contents_getter,
     int frame_tree_node_id) {
   DCHECK_CURRENTLY_ON(content::BrowserThread::UI);
-  io_checker_ = std::make_unique<CheckerOnIO>(
+  sb_checker_ = std::make_unique<CheckerOnSB>(
       weak_factory_.GetWeakPtr(), std::move(delegate_getter),
       web_contents_getter, frame_tree_node_id);
 }
 
 WebApiHandshakeChecker::~WebApiHandshakeChecker() {
   DCHECK_CURRENTLY_ON(content::BrowserThread::UI);
-  content::GetIOThreadTaskRunner({})->DeleteSoon(FROM_HERE,
-                                                 std::move(io_checker_));
+  if (!base::FeatureList::IsEnabled(kSafeBrowsingOnUIThread)) {
+    content::GetIOThreadTaskRunner({})->DeleteSoon(FROM_HERE,
+                                                   std::move(sb_checker_));
+  }
 }
 
 void WebApiHandshakeChecker::Check(const GURL& url, CheckCallback callback) {
   DCHECK_CURRENTLY_ON(content::BrowserThread::UI);
   DCHECK(!check_callback_);
   check_callback_ = std::move(callback);
-  content::GetIOThreadTaskRunner({})->PostTask(
-      FROM_HERE, base::BindOnce(&WebApiHandshakeChecker::CheckerOnIO::Check,
-                                io_checker_->AsWeakPtr(), url));
+  if (base::FeatureList::IsEnabled(kSafeBrowsingOnUIThread)) {
+    sb_checker_->Check(url);
+  } else {
+    content::GetIOThreadTaskRunner({})->PostTask(
+        FROM_HERE, base::BindOnce(&WebApiHandshakeChecker::CheckerOnSB::Check,
+                                  sb_checker_->AsWeakPtr(), url));
+  }
 }
 
-void WebApiHandshakeChecker::OnCompleteCheck(bool slow_check,
-                                             bool proceed,
-                                             bool showed_interstitial) {
+void WebApiHandshakeChecker::OnCompleteCheck(bool proceed) {
   DCHECK_CURRENTLY_ON(content::BrowserThread::UI);
   DCHECK(check_callback_);
